api-gateway/requirements.txt
Flask==3.0.3
Flask-Limiter==3.8.0
requests==2.32.3
python-dotenv==1.0.1
grpcio==1.60.0
grpcio-tools==1.60.0
redis==5.1.1
pybreaker==1.2.0